--- client/platforms/macos_ne/QRCodeReaderBase.cpp
@@ -1,14 +1,0 @@
-#include "QRCodeReaderBase.h"
-
-QRCodeReader::QRCodeReader()
-{
-
-}
-
-QRect QRCodeReader::cameraSize() {
-    return QRect();
-}
-
-void QRCodeReader::startReading() {}
-void QRCodeReader::stopReading() {}
-void QRCodeReader::setCameraSize(QRect) {}
