--- client/platforms/macos_ne/QtAppDelegate.h
@@ -1,7 +1,0 @@
-//#import <UIKit/UIKit.h>
-
-@interface QIOSApplicationDelegate
-@end
-
-@interface QIOSApplicationDelegate (AmneziaVPNDelegate)
-@end
