--- client/platforms/macos_ne/ScreenProtection.swift
@@ -1,10 +1,0 @@
-//import UIKit
-
-public func toggleScreenshots(_ isEnabled: Bool) {
-  
-}
-
-class ScreenProtection {
-
-
-}
