+++ client/ui/notificationhandler.cpp
@@ -5,10 +5,10 @@
 #include <QDebug>
 #include "notificationhandler.h"
 
-#if defined(Q_OS_IOS) && !defined(MACOS_NE)
+#if defined(Q_OS_IOS) || defined(MACOS_NE)
 #  include "platforms/ios/iosnotificationhandler.h"
-#elif defined(MACOS_NE)
-#  include "platforms/macos_ne/iosnotificationhandler.h"
+// #elif defined(MACOS_NE)
+// #  include "platforms/ios/iosnotificationhandler.h"
 #else
 #  include "systemtray_notificationhandler.h"
 #endif
